backend/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.4
uWSGI==2.0.18
Jinja2==2.11.1 
MarkupSafe==1.1.1 
Werkzeug==1.0.1 
click==7.1.1 
itsdangerous==1.1.0 
six==1.14.0
flask_wtf==0.14.3
WTForms==2.2.1
Flask-Cors==3.0.8
flask_sslify==0.1.5
requests==2.23.0